fix(mouth): the clients route shipped a staff name in its public chunk
C4 (#6327) moved the workspace roster tables server-side and wired a build
guard that greps the built artifact. It could not finish `clients/[id]`:
another window's #6329 and #6307 rewrote `page.tsx` and `components/TaxTab.tsx`
while this lane had relocated their contents, and by contract the merged work
wins. So C4 shipped with ONE declared, time-boxed exception —
`ALLOWED_CHUNK_PREFIXES = ["app/(workspace)/clients/"]` — naming this PR as the
one that deletes it. This is that PR, and the list is back to empty.

WHAT WAS WRONG. `TaxTab.tsx` held the five assignable tax consultants as a
module constant, and it is a "use client" module. A literal there is compiled
into `app/(workspace)/clients/%5Bid%5D/page-*.js`, and a Next static chunk is
served from the CDN path with NO session — so an anonymous GET of that chunk
returned a name the owner had excluded from public surfaces. Measured on
production this turn: 2 marker hits in that chunk, on the public domain,
unauthenticated.

WHAT CHANGED. The table is resolved on the server in `page.tsx` (via
`taxConsultants()` in `@/lib/workspace/roster-directory`, the module that stays
in step with backend migration 093's CHECK constraint) and handed down as a
required prop. The 1039-line client body moved to `ClientDetailClient.tsx`
unchanged apart from its signature; `page.tsx` is now a 22-line server
boundary. Only `import type { TaxConsultantOption }` crosses into the client,
and type imports are erased.

The five values are MOVED, never derived. The roster disagrees with two of them
(roster `faysha.tax@` vs the constraint's `faisha.tax@`; roster `tax@` vs
`veronika.tax@`), so deriving them would change what the backend accepts.
C4's `roster-directory.test.ts` pins both spellings and stays green.

MEASURED, not assumed:
- build guard: 522 chunks scanned, 0 skipped by exception (none), no chunk
  carries an excluded member. Independent grep over every .js in
  .next/static/chunks agrees: 0 of 522.
- the route's own chunk: `clients/[id]/page-*.js` 246,027 B, 0 markers.
  The same route on production still serves 2.
- full vitest: 475 files / 5104 tests green. Not a subset — C4's first
  candidate passed eight chosen suites while the full run was red, because a
  guard read a moved file BY PATH. Which is why `page.test.tsx` is renamed and
  repointed here rather than left to break.
- typecheck clean; eslint 0 errors.

The new test asserts EXACT equality between the rendered options and the prop.
Under "contains", re-adding a hardcoded row beside the prop would still pass;
two guilt mutants (ignore the prop; prop plus one extra row) both turn it red,
and the clean tree is green.

NOT fixed here, deliberately: the workspace has no server-side session gate, so
this route's server payload is still sent before `(workspace)/layout.tsx` gates
the viewer. That is a design question about the whole workspace, escalated as
its own work, and widening this guard to cover it would mean encoding the very
thing under review.

Bites: `node scripts/assert-roster-not-in-public-chunks.mjs`, which runs inside
`npm run build` — its OK line now reads "0 skipped by the declared time-boxed
exception (none)", where before this commit it skipped
`app/(workspace)/clients/`.
